Documentation and OOP standard minor fix
