+++ README.md
@@ -1,0 +1,8 @@
+# Word Association
+
+> This is a word association game. Still under active development.
+
+## Tech 
+
+  - Java 8
+  - Javascript ES5 & ES6
