+++ src/wordAssociation/core/components/WordMap.java
@@ -1,12 +1,10 @@
-/**
- * 
- */
-package word_association;
+package wordAssociation.core.components;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 /**
  * Implements the basic framework
@@ -15,12 +13,11 @@ import java.util.List;
  */
 public class WordMap {
 	private List<String> wordList = new ArrayList<String>();
-	private HashMap<String, Word> wordMap = 
-			new HashMap<String, Word>();
-	private ArrayList<String> frontier;
+	private Map<String, Word> wordMap = new HashMap<String, Word>();
+	private List<String> frontier;
 	
 	/**
-	 * Constructor for FriendshipGraph
+	 * Constructor for the WordMap of the word association game
 	 */
 	public WordMap() {
 		this.setFrontier(new ArrayList<String>());
@@ -29,10 +26,11 @@ public class WordMap {
 	
 	/**
 	 * Add a node into the wordMap
+	 * 
 	 * @param data Content on the node
 	 * @param neighbors Neighbor of the node
 	 */
-	public void addVertex(String data, ArrayList<String> neighbors) {
+	public void addVertex(String data, List<String> neighbors) {
 		if (wordList.contains(data)) {
 			System.out.println("Duplicated node added");
 			return;
@@ -49,11 +47,11 @@ public class WordMap {
 		return;
 	}
 
-	public ArrayList<String> getFrontier() {
+	public List<String> getFrontier() {
 		return frontier;
 	}
 
-	public void setFrontier(ArrayList<String> frontier) {
+	public void setFrontier(List<String> frontier) {
 		this.frontier = frontier;
 	}
